Factor out torch-related logic for data loaders
