+++ fedbiomed/common/data/_torch_data_manager.py
@@ -3,52 +3,16 @@ Torch data manager
 """
 
 import math
-from itertools import cycle
-from typing import Union, Tuple, Iterator
+from typing import Union, Tuple
 
 from torch.utils.data import Dataset, Subset, DataLoader
-from torch.utils.data import random_split, SequentialSampler, RandomSampler
+from torch.utils.data import random_split
 
 from fedbiomed.common.exceptions import FedbiomedTorchDataManagerError
 from fedbiomed.common.constants import ErrorNumbers
+from fedbiomed.common.data.loaders import CyclingRandomSampler, CyclingSequentialSampler
 
 from ._sklearn_data_manager import SkLearnDataManager
-
-
-class CyclingSequentialSampler(SequentialSampler):
-    def __iter__(self) -> Iterator[int]:
-        return cycle(super().__iter__())
-
-
-class CyclingRandomSampler:
-    class _Iter:
-        def __init__(self, dataset, **kwargs):
-            self._dataset = dataset
-            self._kwargs = kwargs
-            self._sampler = RandomSampler(self._dataset, **self._kwargs)
-            self._iter = iter(self._sampler)
-            self.num_yielded = 0
-
-        def __next__(self):
-            if self.num_yielded == len(self._sampler):
-                self.reset()
-            self.num_yielded += 1
-            return next(self._iter)
-
-        def reset(self):
-            self.num_yielded = 0
-            self._sampler = RandomSampler(self._dataset, **self._kwargs)
-            self._iter = iter(self._sampler)
-
-    def __init__(self, dataset, **kwargs):
-        self._dataset = dataset
-        self._kwargs = kwargs
-
-    def __len__(self):
-        return len(self._sampler)
-
-    def __iter__(self) -> Iterator[int]:
-        return CyclingRandomSampler._Iter(self._dataset, **self._kwargs)
 
 
 class TorchDataManager(object):
