+++ fedbiomed/common/data/loaders/__init__.py
@@ -39,9 +39,12 @@ At the moment Fed-BioMed provides a custom implementation for numpy-based datase
 [`torch.utils.data.DataLoader`](https://pytorch.org/docs/stable/data.html#torch.utils.data.DataLoader) class.
 """
 from .np_dataloader import NPDataLoader
+from .torch_dataloader import CyclingRandomSampler, CyclingSequentialSampler
 from .utils import _generate_roughly_one_epoch
 
 __all__ = [
     'NPDataLoader',
+    'CyclingRandomSampler',
+    'CyclingSequentialSampler',
     '_generate_roughly_one_epoch'
 ]
